defconfigs: xiaomi: Disable sysrq keys over serial
If the UART console is enabled in the bootloader, this could cause false
triggers, especially if the other side's UART interface has a loose TX
pin (which I've encountered).

Note that sysrq is still enabled because Android init uses it when
shutting down, and it doesn't hurt to keep it enabled for debugging.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>