defconfigs: xiaomi: Disable fortify source
Since GCC-12.0.0 (baremetal), inlining restrictions have been tightened further,
the builtin functions are all freaking out about potential overflows.
Some are being triggered as false positives and would have unintended
behavior if it's attempted to be fixed.
It's better to leave them as it is instead of attempting to fix these
as it can further break things unnecessarily.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>