Defconfig: xiaomi: Enabke CONFIG_INIT_STACK_ALL_ZERO
Init stack is secure as well usefull.
As per logs we have certain uninitialised variables.
so this will be a valid fix for them.

Note that as per comments the usage of this is controversial
But, no adverse effects were noticed on our system
So, I feel it's safe to use on our devices

Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>